ensure merge targets are initialised prior to merge
When merging additional resources/parameters/etc into an existing
Template, the assumption was previously made that the target template
sections had already been initialised.

This assumption was incorrect, as the JSON Decoder does not appear to
initialise maps which are not present in the source json.

To account for this, add a test prior to each section's merge, and
initialise an empty map when necessary just before the merge is
attempted.