ensure Path is output as []interface{} in Map()
A Refish path is stored a []string. However, when outputting a Map(),
compatibility with a Template's internal map[string]interface{}
structure must be maintained. To this end, the Refish path should be
output as []interface{}, not []string. This ensures that repeated
processing of the same Template will be able to walk(...) appropriately.

This is implemented via a hard loop, copying into a new array. Though
there is some potential for some obvious efficiency improvements, this
should be good enough for the expected data.