add --output=parameters option, for outputting parameter values
Add a new command-line flag: --output=parameters|template (defaulting to
"template"), selecting what to output after template processing.

"template" is the same as the previous behaviour.

"parameters" scans the inputs for data relevant to detected Parameters
within the template, and outputs a JSON array suitable for use as the
--parameters command-line option in the create-stack command.

For example, with a template containing:

  {"Parameters": {"Foo": {"Type": "String"}} ...}

And an inputs file containing:

  {"Foo": "Aaa"}

The output with --output=parameters would be:

[{"ParameterKey":"Foo","ParameterValue":"Aaa","UsePreviousValue":false}]