allow multiple --parameters to be specified
allow parameters arguments, such as --parameters=./someparams.json to be
specified multiple times.

Each time will be "attached" to the inputs as an Override, using a
FallbackMap, so later specifications will take priority over earlier
ones.

As part of this change, the "inputs" system is renamed and simplified to
"fallbackmap".