allow top-level maps to be excluded
previously, top-level maps could not be excluded by Rules. This was
problematic as some functions pass the path down unmodified, treating
their arguments as top-level during processing.

The mechanics of this issue were that Rules would return a nil key to
indicate that they were to be excluded (this was reasonable and
intuitive: where do I put this returned data? nowhere.) As top-level
objects already had no key, this ambiguity was resolved as "you cannot
exclude the top-level". While this was also perfectly reasonable to say,
it meant that functions which process their arguments with an unmodified
path had no way of indicating exclusion.

Instead, use the boolean "true" to indicate exclusion. The mnemonic here
is "skip = true" to skip.