remove "component" functionality
previously, functionality existed to "automatically" include external
"components", based on the presence of a "Ref". This functionality was
inconsistent with the behaviour of the rest of the processor, as it
involves potentially modifying *any* part of the output template.

This makes it hard to reason about things such as "Is this Ref a
Parameter?" as we cannot know in advance whether or not a Parameter will
be added later on.

Now that Fn::IncludeFile has been added, along with Fn::MergeDeep, the
"Component" pattern is redundant. While we lose the "automatic" nature
of Components, the consistency and simplified reasoning should make up
for it.

This change *does* break older templates which rely on "component"
functionality. As we do not have a "stable" release yet, that is deemed
acceptable.