Evaluate input parameters on-demand
Previously, input parameters were loaded all-at-once, as soon as they
were specified. This meant that complex input assignments would not
account for overrides. To avoid potential issues surrounding this, the
input parameters went completely unprocessed prior to lookup. This meant
that derived paths (eg: paths which would only be visible after the
processing of rules such as "Fn::Merge") were completely inaccessible
within the templates.

This was extremely inconvenient, as it limited possibilities such as
environment-based parameter switching to essentially glorified flat
files. This was partially worked-around by running output "credentials"
files through the parser prior to output, but this made the values at
different processing stages inconsistent.

Ideally, a "generated" path should be just-as accessible as a hard-coded
one.

To allow this, we keep track of which paths have already been accessed.
And previously-unaccessed path gets run through the processor on-demand.
Because the processor only runs when the path is specifically requested
(presumably, by another stage of the processor), we can be relatively
certain that *all* overrides will be accounted for during processing.

This has almost certainly introduced a potential for an infinite loop /
stack-too-deep error.

Other caveats regarding input variables still exist. For example,
requesting a "higher-level" path will not automatically merge-in
those more-precise paths which have been specified via an override, as
determining precisely how to perform this merge is complex enough that
it cannot accurately be automatically determined. However, the issue of
all inputs needing to be specified in a "hard-coded" manner should now
be eliminated.